Changes to Loss Functions
# Loss Accumulation
Losses can now be accumlated over multiple iterations before applying back-prop.

# Loss Logging
Additionally. losses that are logged and shown in tensorboard are the raw losses, independent of their loss_weight. Total GAN loss is added to better understand how the network is learning over time.

# Zero Losses
Loss weights can be set to 0 and the training step will skip them entirely.

# GradScaler Fix
In the prior implementation, a GradScaler was initalised repeatedly in the optimize_parameters function. The .update() function at the end of the code was redundant because of this. Additionally, scales were shared across the generator and discriminator. Each model now has their own GradScaler which correctly updates over time.